refactor(video): move encoder declarations to header (#2185)
